Added product delete feature

+++ src/main/java/com/example/product_service_12052024/models/Product.java
@@ -9,10 +9,10 @@ import lombok.Setter;
 public class Product {
    
    
-   private int id;
+   private Long id;
    private String title;
    private String description;
-   private double price;
+   private Double price;
    private String imageUrl;
    private Category category;
